Fix hash key of pushed layer, when using cache
Relates to https://github.com/GoogleContainerTools/kaniko/issues/899
When Kaniko is pushing a layer at the end of a Dockerfile, if previous line have COPY inside, the computeKey is not compute well.
It happen cause the cachingcopycommand (use for cached layer) doesn't support the FilesUsedFromContext method, it was using the base_command one.
I've fixed it by doing a copy paste of the copycommand, and by adding the buildcontext parameter.
I think It will need a rework to refactor it, but I don't know how to use CacheCopy instead of BaseCommand on line 169 (it crash at runtime)